created checkIsFavorited in the helpers
