+++ pages/browse/shows/[slug].vue
@@ -7,7 +7,8 @@ import ShareIcon from '~/components/icons/ShareIcon.vue'
 
 import {
   deleteFavorite,
-  saveFavorite
+  saveFavorite,
+  checkIsFavorited,
 } from '~/utilities/helpers'
 
 const config = useRuntimeConfig()
@@ -40,7 +41,7 @@ const handleStar = () => {
   const show = {
     slug: route.params.slug,
   }
-  if(isFavorited.value){
+  if (isFavorited.value) {
     deleteFavorite(show, 'show')
     isFavorited.value = false
   } else {
@@ -61,22 +62,8 @@ watch(show, () => {
 })
 
 // if user is logged in, check if item is already favorited
-const client = useSupabaseClient()
-const isFavorited = ref(false)
+const isFavorited = ref(await checkIsFavorited(route.params.slug))
 const user = useCurrentUser()
-if (user.value) {
-  const { data, error } = await client
-    .from('favorited')
-    .select('*')
-    .eq('uid', user.value.id)
-    .eq('media_slug', route.params.slug)
-    if(data?.length > 0){
-      isFavorited.value = true
-    }
-    if(error){
-      console.log('favorited items error', error)
-    }
-}
 </script>
 
 <template>
@@ -117,7 +104,9 @@ if (user.value) {
       class="flex justify-content-center align-items-center gap-2 mt-2 mb-4"
     >
       <Button v-if="user" rounded text plain @click="handleStar">
-        <template #icon> <StarIcon :active="isFavorited" class="w-2rem" /></template>
+        <template #icon>
+          <StarIcon :active="isFavorited" class="w-2rem"
+        /></template>
       </Button>
 
       <Button
